views/examples: Fix the format string of size_t var in RadioButton example.
Cast from size_t to int.

TEST=compile views_examples target, it should work without errors.
R=sky@chromium.org

Review URL: http://codereview.chromium.org/8252009

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@105584 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: e7d05f9c98a95e6b388ec3a7d817627d1f4fb821

+++ views/examples/radio_button_example.cc
@@ -26,7 +26,8 @@ void RadioButtonExample::CreateExampleView(views::View* container) {
   int group = 1;
   for (size_t i = 0; i < arraysize(radio_buttons_); ++i) {
     radio_buttons_[i] = new views::RadioButton(
-        UTF8ToUTF16(base::StringPrintf("Radio %d in group %d", i + 1, group)),
+        UTF8ToUTF16(base::StringPrintf(
+            "Radio %d in group %d", static_cast<int>(i) + 1, group)),
         group);
     radio_buttons_[i]->set_listener(this);
   }
